Move classes to proper packages
