+++ platform/base/core/src/main/java/com/peregrine/reference/impl/ReferenceListerService.java
@@ -1,4 +1,4 @@
-package com.peregrine.render.impl;
+package com.peregrine.reference.impl;
 
 /*-
  * #%L
